+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Please use [this](https://docs.gitlab.com/ee/development/changelog.html) document as guidelines to keep a changelog.
 
+## [unreleased]
+
+### Added
+
+* ipfs: Added `ipfs` datasource to download JSON and binary data from IPFS.
+
 ## 4.1.2 - 2022-01-27
 
 ### Added